feat: make FlatList as a shared component
