+++ src/components/flatlist/index.ts
@@ -1,0 +1,1 @@
+export { default as FlatList } from './flat-list'
